Removing info specific code from home component
