+++ src/components/home.tsx
@@ -7,10 +7,8 @@ import MenuItem from '@material-ui/core/MenuItem';
 import Email from './email';
 import Info from './info';
 import Links from './links';
-import makeRequest from '../lib/makeRequest';
 import isValidUrl from '../lib/isValidUrl';
 import './home.css';
-import { string } from 'prop-types';
 
 const StyledTextField = withStyles({
     root: {
@@ -34,20 +32,7 @@ const LINKS = 'GET_LINKS';
 const INFO = 'GET_INFORMATION';
 const EMAILS = 'GET_EMAILS';
 
-const getHeaderMap = (request: XMLHttpRequest): Map<string, string> => {
-    const headers = request.getAllResponseHeaders();
-    const arr = headers.trim().split(/[\r\n]+/);
-    const headerMap = new Map();
-
-    arr.forEach(line => {
-      var parts = line.split(': ');
-      var header = parts.shift();
-      var value = parts.join(': ');
-      headerMap.set(header, value);
-    });
-
-    return headerMap;
-}
+type HomeProps = {};
 
 type HomeState = {
     option: string,
@@ -55,8 +40,6 @@ type HomeState = {
     info: Map<string, string>,
     submit: boolean
 };
-
-type HomeProps = {};
 
 export default class Home extends React.Component<HomeProps, HomeState> {
     constructor(props: HomeProps) {
@@ -74,23 +57,7 @@ export default class Home extends React.Component<HomeProps, HomeState> {
             alert('Invalid URL');
             return;
         }
-        switch (this.state.option) {
-            case LINKS:
-            case EMAILS:
-                this.setState({'submit': true});
-                break;
-            case INFO:
-                makeRequest('GET', this.state.url)
-                    .then(response => {
-                        const headers = getHeaderMap(response);
-                        this.setState({'info': headers});
-                        this.setState({'submit': true});
-                    })
-                    .catch(error => {
-                        console.error(error);
-                    })
-                break;
-        }
+        this.setState({'submit': true});
     }
     
     handleTextChange(event: React.ChangeEvent<HTMLInputElement>) {
@@ -127,7 +94,7 @@ export default class Home extends React.Component<HomeProps, HomeState> {
         }
         switch (this.state.option) {
             case INFO:
-                return <Info info={this.state.info}/>;
+                return <Info url={this.state.url}/>;
             case LINKS:
                 return <Links url={this.state.url}/>;
             case EMAILS:
